put logs in its own component
